+++ src/main/java/com/iota/iri/SignedFiles.java
@@ -13,16 +13,16 @@ import java.util.Arrays;
 public class SignedFiles {
 
     public static boolean isFileSignatureValid(String filename, String signatureFilename, String publicKey, int depth, int index) throws IOException {
-        int[] signature = digestFile(filename, SpongeFactory.create(SpongeFactory.Mode.KERL));
+        byte[] signature = digestFile(filename, SpongeFactory.create(SpongeFactory.Mode.KERL));
         return validateSignature(signatureFilename, publicKey, depth, index, signature);
     }
 
-    private static boolean validateSignature(String signatureFilename, String publicKey, int depth, int index, int[] digest) throws IOException {
+    private static boolean validateSignature(String signatureFilename, String publicKey, int depth, int index, byte[] digest) throws IOException {
         //validate signature
         SpongeFactory.Mode mode = SpongeFactory.Mode.CURLP81;
-        int[] digests = new int[0];
-        int[] bundle = ISS.normalizedBundle(digest);
-        int[] root;
+        byte[] digests = new byte[0];
+        byte[] bundle = ISS.normalizedBundle(digest);
+        byte[] root;
         int i;
 
         try (InputStream inputStream = SignedFiles.class.getResourceAsStream(signatureFilename);
@@ -31,33 +31,33 @@ public class SignedFiles {
 
             String line;
             for (i = 0; i < 3 && (line = reader.readLine()) != null; i++) {
-                int[] lineTrits = Converter.allocateTritsForTrytes(line.length());
+                byte[] lineTrits = Converter.allocateTritsForTrytes(line.length());
                 Converter.trits(line, lineTrits, 0);
-                int[] normalizedBundleFragment = Arrays.copyOfRange(bundle, i * ISS.NORMALIZED_FRAGMENT_LENGTH, (i + 1) * ISS.NORMALIZED_FRAGMENT_LENGTH);
-                int[] issDigest = ISS.digest(mode, normalizedBundleFragment, lineTrits);
+                byte[] normalizedBundleFragment = Arrays.copyOfRange(bundle, i * ISS.NORMALIZED_FRAGMENT_LENGTH, (i + 1) * ISS.NORMALIZED_FRAGMENT_LENGTH);
+                byte[] issDigest = ISS.digest(mode, normalizedBundleFragment, lineTrits);
                 digests = ArrayUtils.addAll(digests, issDigest);
             }
 
             if ((line = reader.readLine()) != null) {
-                int[] lineTrits = Converter.allocateTritsForTrytes(line.length());
+                byte[] lineTrits = Converter.allocateTritsForTrytes(line.length());
                 Converter.trits(line, lineTrits, 0);
                 root = ISS.getMerkleRoot(mode, ISS.address(mode, digests), lineTrits, 0, index, depth);
             } else {
                 root = ISS.address(mode, digests);
             }
 
-            int[] pubkeyTrits = Converter.allocateTritsForTrytes(publicKey.length());
+            byte[] pubkeyTrits = Converter.allocateTritsForTrytes(publicKey.length());
             Converter.trits(publicKey, pubkeyTrits, 0);
             return Arrays.equals(pubkeyTrits, root); // valid
         }
     }
 
-    private static int[] digestFile(String filename, Sponge curl) throws IOException {
+    private static byte[] digestFile(String filename, Sponge curl) throws IOException {
         try (InputStream inputStream = SignedFiles.class.getResourceAsStream(filename);
              BufferedReader reader = new BufferedReader((inputStream == null)
                  ? new FileReader(filename) : new InputStreamReader(inputStream))) {
 
-            int[] buffer = new int[Curl.HASH_LENGTH * 3];
+            byte[] buffer = new byte[Curl.HASH_LENGTH * 3];
 
             reader.lines().forEach(line -> {
                 String trytes = Converter.asciiToTrytes(line); // can return a null
@@ -66,10 +66,10 @@ public class SignedFiles {
                 }
                 Converter.trits(trytes, buffer, 0);
                 curl.absorb(buffer, 0, buffer.length);
-                Arrays.fill(buffer, 0);
+                Arrays.fill(buffer, (byte) 0);
             });
 
-            int[] signature = new int[Curl.HASH_LENGTH];
+            byte[] signature = new byte[Curl.HASH_LENGTH];
             curl.squeeze(signature, 0, Curl.HASH_LENGTH);
             return signature;
         } catch (UncheckedIOException e) {
